requirements.txt
gitpython>=3.1.40
javalang>=0.13.0
typer[all]>=0.9.0
rich>=13.7.0
pyautogen>=0.2.28  # For AutoGen evaluation agent 